Conflicts:
	BikeWale.UI/BikeWale.csproj
	BikeWale.UI/new/versions.aspx

conflict with release in csproj file Resolved (for tnc.html file)

+++ BikeWale.UI/BindViewModels/Controls/BindSimilarVideos.cs
@@ -1,6 +1,5 @@
 ﻿using Bikewale.Cache.Core;
 using Bikewale.Cache.Videos;
-using Bikewale.DAL.Videos;
 using Bikewale.Entities.Videos;
 using Bikewale.Interfaces.Cache.Core;
 using Bikewale.Interfaces.Videos;
@@ -18,7 +17,7 @@ namespace Bikewale.BindViewModels.Controls
 {
     public class BindSimilarVideos
     {
-        public uint TotalRecords { get; set; }
+        public ushort TotalRecords { get; set; }
         public int FetchedRecordsCount { get; set; }
         public EnumVideosCategory CategoryId { get; set; }
         public uint VideoBasicId { get; set; }
